Fixing trash comits

+++ jama/src/main/java/com/cibertec/jama/repositories/menu/MenuCategoryRepository.java
@@ -3,5 +3,5 @@ package com.cibertec.jama.repositories.menu;
 import com.cibertec.jama.entities.menu.MenuCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuCategoryRepository extends JpaRepository<MenuCategory, Long> {
+public interface MenuCategoryRepository extends JpaRepository<MenuCategory, Integer> {
 }
