+++ jama/src/main/java/com/cibertec/jama/repositories/menu/MenuSkuRepository.java
@@ -1,8 +1,13 @@
 package com.cibertec.jama.repositories.menu;
 
+import com.cibertec.jama.entities.menu.Menu;
 import com.cibertec.jama.entities.menu.MenuSku;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuSkuRepository  extends JpaRepository<MenuSku, Long> {
+import java.util.List;
+
+public interface MenuSkuRepository  extends JpaRepository<MenuSku, Integer> {
+
+
 }
