+++ jama/src/main/java/com/cibertec/jama/repositories/menu/MenuTypeRepository.java
@@ -4,5 +4,5 @@ import com.cibertec.jama.entities.menu.MenuType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuTypeRepository extends JpaRepository<MenuType, Long> {
+public interface MenuTypeRepository extends JpaRepository<MenuType, Integer> {
 }
